[refactor] Add ex handler
related to #40
